moving user nav helpers into sso utils
