+++ src/platform/user/tests/authentication/utilities.unit.spec.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 
-import * as ssoUtils from 'platform/utilities/sso';
+import * as forceAuth from 'platform/utilities/sso/forceAuth';
 
 import {
   login,
@@ -71,7 +71,7 @@ describe('authentication URL helpers', () => {
   });
 
   it('should redirect for login v1 with force auth', () => {
-    const stub = sinon.stub(ssoUtils, 'getForceAuth').callsFake(() => true);
+    const stub = sinon.stub(forceAuth, 'getForceAuth').callsFake(() => true);
     login('idme', 'v1');
     stub.restore();
     expect(global.window.location).to.include(
@@ -80,7 +80,7 @@ describe('authentication URL helpers', () => {
   });
 
   it('should redirect for login v1 with application and force auth', () => {
-    const stub = sinon.stub(ssoUtils, 'getForceAuth').callsFake(() => true);
+    const stub = sinon.stub(forceAuth, 'getForceAuth').callsFake(() => true);
     login('idme', 'v1', 'my-app');
     stub.restore();
     expect(global.window.location).to.include(
