+++ src/platform/utilities/sso/index.js
@@ -1,9 +1,11 @@
 import moment from 'moment';
 import environment from 'platform/utilities/environment';
 import localStorage from '../storage/localStorage';
-import { hasSessionSSO } from '../../user/profile/utilities';
+import { hasSession, hasSessionSSO } from '../../user/profile/utilities';
+import { autoLogin, autoLogout } from 'platform/user/authentication/utilities';
 import mockKeepAlive from './mockKeepAliveSSO';
 import liveKeepAlive from './keepAliveSSO';
+import { getForceAuth } from './forceAuth';
 
 const keepAlive = environment.isLocalhost() ? mockKeepAlive : liveKeepAlive;
 const keepAliveThreshold = 5 * 60 * 1000; // 5 minutes, in milliseconds
@@ -24,6 +26,23 @@ export async function ssoKeepAliveSession() {
     localStorage.removeItem('hasSessionSSO');
   }
 }
+
+export async function checkAutoSession(callback) {
+  await ssoKeepAliveSession();
+  if (hasSession() && hasSessionSSO() === false) {
+    // explicitly check to see if the SSOe session is false, as it could also
+    // be null if we failed to get a response from the SSOe server, in which
+    // case we don't want to logout the user because we don't know
+    autoLogout();
+  } else if (!hasSession() && hasSessionSSO() && !getForceAuth()) {
+    autoLogin();
+  }
+
+  if (callback) {
+    callback();
+  }
+}
+
 export function checkAndUpdateSSOeSession() {
   if (hasSessionSSO()) {
     const sessionExpiration = localStorage.getItem('sessionExpirationSSO');
@@ -34,20 +53,3 @@ export function checkAndUpdateSSOeSession() {
     }
   }
 }
-
-/*
- * When set, the `requiresForceAuth` indicates that the current unauthenticated
- * user will have to "force" authentication against SSOe in order to establish a
- * session.
- */
-export const getForceAuth = () => {
-  localStorage.getItem('requiresForceAuth');
-};
-
-export const setForceAuth = () => {
-  localStorage.setItem('requiresForceAuth', true);
-};
-
-export const removeForceAuth = () => {
-  localStorage.removeItem('requiresForceAuth');
-};
