+++ src/platform/utilities/tests/sso.unit.spec.js
@@ -1,8 +1,13 @@
 import { expect } from 'chai';
+import sinon from 'sinon';
 
 import { mockFetch, resetFetch } from 'platform/testing/unit/helpers';
 import localStorage from 'platform/utilities/storage/localStorage';
-import { checkAndUpdateSSOeSession } from '../sso';
+import * as authUtils from 'platform/user/authentication/utilities';
+import * as profUtils from 'platform/user/profile/utilities';
+
+import { checkAutoSession, checkAndUpdateSSOeSession } from '../sso';
+import * as forceAuth from '../sso/forceAuth';
 
 function setKeepAliveResponse(stub, sessionTimeout = 0) {
   const response = new Response();
@@ -20,6 +25,85 @@ function setKeepAliveResponse(stub, sessionTimeout = 0) {
 }
 
 const SSO_SESSION_TIMEOUT = 900; // seconds
+
+describe('checkAutoSession', () => {
+  it('should auto logout if user is logged in and they do not have a SSOe session', async () => {
+    const sandbox = sinon.createSandbox();
+    sandbox.stub(profUtils, 'hasSession').returns(true);
+    sandbox.stub(profUtils, 'hasSessionSSO').returns(false);
+    sandbox.stub(forceAuth, 'getForceAuth').returns(undefined);
+    const auto = sandbox.stub(authUtils, 'autoLogout');
+    await checkAutoSession();
+    sandbox.restore();
+    sinon.assert.calledOnce(auto);
+  });
+
+  it('should not auto logout if user is logged in and they have a SSOe session', async () => {
+    const sandbox = sinon.createSandbox();
+    sandbox.stub(profUtils, 'hasSession').returns(true);
+    sandbox.stub(profUtils, 'hasSessionSSO').returns(true);
+    sandbox.stub(forceAuth, 'getForceAuth').returns(undefined);
+    const auto = sandbox.stub(authUtils, 'autoLogout');
+    await checkAutoSession();
+    sandbox.restore();
+    sinon.assert.notCalled(auto);
+  });
+
+  it('should not auto logout if user is logged in and we dont know if they have a SSOe session', async () => {
+    const sandbox = sinon.createSandbox();
+    sandbox.stub(profUtils, 'hasSession').returns(true);
+    sandbox.stub(profUtils, 'hasSessionSSO').returns(null);
+    sandbox.stub(forceAuth, 'getForceAuth').returns(undefined);
+    const auto = sandbox.stub(authUtils, 'autoLogout');
+    await checkAutoSession();
+    sandbox.restore();
+    sinon.assert.notCalled(auto);
+  });
+
+  it('should auto login if user is logged out, they have a SSOe session and dont need to force auth', async () => {
+    const sandbox = sinon.createSandbox();
+    sandbox.stub(profUtils, 'hasSession').returns(false);
+    sandbox.stub(profUtils, 'hasSessionSSO').returns(true);
+    sandbox.stub(forceAuth, 'getForceAuth').returns(undefined);
+    const auto = sandbox.stub(authUtils, 'autoLogin');
+    await checkAutoSession();
+    sandbox.restore();
+    sinon.assert.calledOnce(auto);
+  });
+
+  it('should not auto login if user is logged out, they dont have a SSOe session and dont need to force auth', async () => {
+    const sandbox = sinon.createSandbox();
+    sandbox.stub(profUtils, 'hasSession').returns(false);
+    sandbox.stub(profUtils, 'hasSessionSSO').returns(false);
+    sandbox.stub(forceAuth, 'getForceAuth').returns(undefined);
+    const auto = sandbox.stub(authUtils, 'autoLogin');
+    await checkAutoSession();
+    sandbox.restore();
+    sinon.assert.notCalled(auto);
+  });
+
+  it('should not auto login if user is logged out, they have a SSOe session and need to force auth', async () => {
+    const sandbox = sinon.createSandbox();
+    sandbox.stub(profUtils, 'hasSession').returns(false);
+    sandbox.stub(profUtils, 'hasSessionSSO').returns(true);
+    sandbox.stub(forceAuth, 'getForceAuth').returns(true);
+    const auto = sandbox.stub(authUtils, 'autoLogin');
+    await checkAutoSession();
+    sandbox.restore();
+    sinon.assert.notCalled(auto);
+  });
+
+  it('should invoke callback if provided', async () => {
+    const sandbox = sinon.createSandbox();
+    sandbox.stub(profUtils, 'hasSession').returns(false);
+    sandbox.stub(profUtils, 'hasSessionSSO').returns(false);
+    sandbox.stub(forceAuth, 'getForceAuth').returns(undefined);
+    const callback = sandbox.spy();
+    await checkAutoSession(callback);
+    sandbox.restore();
+    sinon.assert.calledOnce(callback);
+  });
+});
 
 describe('checkAndUpdateSSOeSession', () => {
   it('should should do nothing if there is not SSO session active', () => {
